Temperture model

+++ Atmel Studio/Diode Temp Sensor/LCDWrite.h
@@ -13,9 +13,9 @@
 #include <stdio.h>
 #include <avr/io.h>
 #include "lcd.h"
+#include "TempModel.h"
 
-
-void extern lcd_Write( uint16_t adc_reading_lm35, uint16_t adc_reading_diode, uint8_t temperture_lm35, uint8_t temperture_diode );
+void extern lcd_Write( uint16_t adc_reading_lm35, uint16_t adc_reading_diode, uint16_t temperture_lm35, uint16_t temperture_diode);
